lab2/Galanin_Pavel: parseInt to parseFloat

+++ Reports/Galanin_Pavel/lab2/_myPublicFolder/task-3/task-3.js
@@ -14,7 +14,7 @@ while (1) {
         alert("Input - not a number\n(ввод - не число)");
         break;
     }
-    x = parseInt(x);
+    x = parseFloat(x);
     //task
     alert(`Fibonacci(${x}) = ${getFibonacci(x)}`);
 } //end infinity while
